+++ pkgs/yaml_edit/test/string_test.dart
@@ -9,6 +9,8 @@ import 'package:yaml_edit/yaml_edit.dart';
 final _testStrings = [
   "this is a fairly' long string with\nline breaks",
   'whitespace\n after line breaks',
+  'whitespace\n \nbetween line breaks',
+  '\n line break at the start',
   'word',
   'foo bar',
   'foo\nbar',
@@ -21,7 +23,7 @@ final _testStrings = [
 final _scalarStyles = [
   ScalarStyle.ANY,
   ScalarStyle.DOUBLE_QUOTED,
-  //ScalarStyle.FOLDED, // TODO: Fix this test case!
+  ScalarStyle.FOLDED,
   ScalarStyle.LITERAL,
   ScalarStyle.PLAIN,
   ScalarStyle.SINGLE_QUOTED,